refactor(imports): removing proto/index file
this removes a double export
