+++ src/api/client.ts
@@ -9,14 +9,14 @@ import {
     PingRequest,
     PingResponse,
     SubscribeStatesRequest,
-    voidMessage,
-} from './protobuf';
+} from './protobuf/api';
 import { ReadData } from './connection';
 import { Observable, of, Subscription } from 'rxjs';
 import { MessageTypes } from './requestResponseMatching';
 import { filter, switchMap, take, tap } from 'rxjs/operators';
 import { Reader } from 'protobufjs/minimal';
 import { EspSocket } from './espSocket';
+import { voidMessage } from './protobuf/api_options';
 
 export interface Decoder<T> {
     decode: (reader: Reader, length?: number) => T;
