--- src/api/protobuf/index.ts
@@ -1,3 +1,0 @@
-export * from './api';
-export * from './api_options';
-export * from './google_descriptor';
